Reorganised so i can build an exe from the python
