--- NewAnalysis/curateFolder.py
@@ -1,45 +1,0 @@
-import os, sys
-import curateCamera as cc
-import curateEngine as ce
-import configparser as cfg
-import ReadUFOCapXML
-
-if __name__ == '__main__':
-    path = sys.argv[1]
-    config=cfg.ConfigParser()
-
-    # if the argument is an ini file, read from it.
-    # otherwise look for a file testing.ini in the target folder
-    if path[-4:]=='.ini':
-        cfgfile=path
-        config.read(cfgfile)
-        path=config['camera']['localfolder']
-    else:
-        cfgfile=os.path.join(path, 'testing.ini')
-        config.read(cfgfile)
-
-    cc.badfilepath=config['cleaning']['badfolder']
-    logfilepath=cc.badfilepath
-    ce.MAXRMS=float(config['cleaning']['maxrms'])
-    ce.MINLEN=int(config['cleaning']['minlen'])
-    ce.MAXLEN=int(config['cleaning']['maxlen'])
-    ce.MAXBRI=int(config['cleaning']['maxbri'])
-    ce.MAXOBJS=int(config['cleaning']['maxobjs'])
-    print(config['cleaning']['debug'])
-    if config['cleaning']['debug'] in ['True', 'TRUE','true']:
-        ce.debug=True
-    else:
-        ce.debug=False
-    if config['cleaning']['movefiles'] in ['True', 'TRUE','true']:
-        cc.movfiles=True
-    else:
-        cc.movfiles=False
-
-    if config['cleaning']['useSubfolders'] in ['True', 'TRUE','true']:
-        cc.useSubfolders=True
-    else:
-        cc.useSubfolders=False
-
-    print('Processing '+path+'; cc.movefiles='+str(cc.movfiles))
-
-    cc.ProcessADay(path, '*', cc.badfilepath, logfilepath)
